Replace the "review" page
Now follows designs from https://www.figma.com/file/j8OounnzVVLno1EEhnN6Wp/Untitled?node-id=0%3A1
