+++ redwood/run_tests.sh
@@ -3,7 +3,10 @@
 set -eo pipefail
 
 yarn install --frozen-lockfile
-yarn rw tsc
-yarn rw lint
+yarn rw tsc &
+yarn rw lint&
+wait
+
 yarn rw test --watch=false --ci
 yarn rw check
+wait
