+++ redwood/web/src/Routes.tsx
@@ -22,9 +22,9 @@ const Routes = () => {
           <Route path="/sign-up/connect" page={SignUpConnectWalletPage} name="signUpConnectWallet" />
           <Route path="/sign-up/allow-camera" page={SignUpAllowCameraPage} name="signUpAllowCamera" />
           <Route path="/sign-up/record" page={SignUpRecordPage} name="signUpRecord" />
-          <Route path="/sign-up/presubmit" page={SignUpPresubmitPage} name="signUpPresubmit" />
+          <Route path="/sign-up/review" page={SignUpReviewPage} name="signUpReview" />
+          <Route path="/sign-up/submitted" page={SignUpSubmittedPage} name="signUpSubmitted" />
         </Set>
-        <Route path="/sign-up/submitted" page={SignUpSubmittedPage} name="signUpSubmitted" />
       </Set>
     </Router>
   )
