--- redwood/web/src/pages/SignUp/VideoField.tsx
@@ -1,222 +1,0 @@
-import {
-  Box,
-  Button,
-  ButtonGroup,
-  Link,
-  ListItem,
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalHeader,
-  ModalOverlay,
-  OrderedList,
-  ScaleFade,
-  Stack,
-  Text,
-  useDisclosure,
-  UseDisclosureReturn,
-} from '@chakra-ui/react'
-import {useState} from 'react'
-import ReactPlayer from 'react-player'
-import Webcam from 'react-webcam'
-import {VideoBox} from 'src/components/SquareBox'
-import {useAppDispatch, useAppSelector} from 'src/state/store'
-import {signUpSlice} from 'src/state/signUpSlice'
-import {useFilePicker} from 'use-file-picker'
-
-const VideoModal = (props: {modalCtrl: UseDisclosureReturn}) => {
-  const dispatch = useAppDispatch()
-
-  const [webcamActive, setwebcamActive] = useState<boolean>(false)
-  const [recording, setRecording] = useState(false)
-
-  const [candidateVid, setCandidateVid] = useState<string | undefined>()
-
-  const [openFileSelector, {filesContent}] = useFilePicker({
-    readAs: 'DataURL',
-    accept: 'video/*',
-    multiple: true,
-    limitFilesConfig: {max: 1},
-  })
-
-  React.useEffect(() => {
-    if (filesContent.length < 1) return
-    ;(async () => setCandidateVid(filesContent[0].content))()
-  }, [filesContent])
-
-  const webcamRef = React.useRef<Webcam>(null)
-  const mediaRecorderRef = React.useRef<MediaRecorder>(null)
-
-  const handleStartCaptureClick = React.useCallback(() => {
-    setRecording(true)
-    // @ts-expect-error TODO: why are we assigning to a supposedly readonly ref
-    // here? Just copied the example from
-    // https://codepen.io/mozmorris/pen/yLYKzyp?editors=0010
-    mediaRecorderRef.current = new MediaRecorder(webcamRef.current.stream, {
-      mimeType: 'video/webm',
-    })
-    mediaRecorderRef.current.addEventListener('dataavailable', ({data}) => {
-      // Create a copy of the blob with the mime type set to 'video/webm'.
-      // Without this at least on MacOS/Chrome I get a mime type of
-      // video/x-matroska;codecs=avc1,opus, which Infura refuses to accept
-      // as a valid mime-type when uploading the video.
-      const video = data.slice(0, data.size, 'video/webm')
-      setCandidateVid(URL.createObjectURL(video))
-    })
-    mediaRecorderRef.current.start()
-  }, [webcamRef, setRecording, mediaRecorderRef])
-
-  const handleStopCaptureClick = React.useCallback(() => {
-    mediaRecorderRef.current?.stop()
-    setRecording(false)
-    setwebcamActive(false)
-  }, [mediaRecorderRef, webcamRef, setRecording])
-
-  const saveVideo = () => {
-    if (candidateVid == null) return
-
-    dispatch(signUpSlice.actions.setVideo(candidateVid))
-    setCandidateVid(undefined)
-    props.modalCtrl.onClose()
-  }
-
-  return (
-    <Modal
-      isOpen={props.modalCtrl.isOpen}
-      onClose={props.modalCtrl.onClose}
-      size="xl"
-    >
-      <ModalOverlay />
-      <ModalContent>
-        <ModalHeader>Record a Video</ModalHeader>
-        <ModalBody>
-          <Stack spacing="8" align="center">
-            <Box>
-              <Text mb="2">
-                Recording a video makes it harder for bots to get into the
-                registry. In your video, you should:
-              </Text>
-              <OrderedList fontSize="sm">
-                <ListItem>Make sure your face is fully visible.</ListItem>
-                <ListItem>
-                  Raise your right hand, and repeat the following:
-                  <br />
-                  <strong>
-                    "I solemnly swear this is my first time registering a Zorro
-                    account"
-                  </strong>
-                </ListItem>
-              </OrderedList>
-            </Box>
-            {webcamActive && (
-              <>
-                <Box
-                  borderRadius="lg"
-                  overflow="hidden"
-                  bgColor="gray.600"
-                  width="100%"
-                >
-                  <Webcam
-                    videoConstraints={{facingMode: 'user'}}
-                    audio
-                    muted
-                    ref={webcamRef}
-                  />
-                </Box>
-                <ButtonGroup justify="center" pb="4">
-                  {recording ? (
-                    <Button
-                      colorScheme="green"
-                      onClick={handleStopCaptureClick}
-                    >
-                      Finish Recording
-                    </Button>
-                  ) : (
-                    <>
-                      <Button
-                        colorScheme="red"
-                        onClick={handleStartCaptureClick}
-                      >
-                        Start Recording
-                      </Button>
-                      <Button onClick={() => setwebcamActive(false)}>
-                        Cancel
-                      </Button>
-                    </>
-                  )}
-                </ButtonGroup>
-              </>
-            )}
-            {candidateVid && !webcamActive && (
-              <ScaleFade in={!!candidateVid && !webcamActive}>
-                <Stack spacing="8" align="center">
-                  <Box borderRadius="md" overflow="hidden" shadow="md">
-                    <ReactPlayer
-                      url={candidateVid}
-                      controls
-                      width="100%"
-                      height="auto"
-                    />
-                  </Box>
-                  <ButtonGroup justify="center" pb="4">
-                    <Button colorScheme="blue" onClick={saveVideo}>
-                      Use This
-                    </Button>
-                    <Button onClick={() => setCandidateVid(undefined)}>
-                      Choose Another
-                    </Button>
-                  </ButtonGroup>
-                </Stack>
-              </ScaleFade>
-            )}
-            {!candidateVid && !webcamActive && (
-              <ButtonGroup justify="center" pb="4">
-                <Button onClick={openFileSelector}>Upload Video</Button>
-                <Button
-                  onClick={() => setwebcamActive(true)}
-                  colorScheme="purple"
-                >
-                  Use Webcam
-                </Button>
-              </ButtonGroup>
-            )}
-          </Stack>
-        </ModalBody>
-        <ModalCloseButton />
-      </ModalContent>
-    </Modal>
-  )
-}
-
-const VideoField: React.FC<{readOnly?: boolean}> = (props) => {
-  const {readOnly = false} = props
-  const modalControl = useDisclosure()
-
-  const video = useAppSelector((state) => state.signUp.video)
-
-  return (
-    <>
-      <VideoModal modalCtrl={modalControl} />
-      {video ? (
-        <Stack>
-          <VideoBox video={video} width="36" shadow="lg" />
-          {!readOnly && (
-            <Link
-              as="button"
-              type="button"
-              variant="btn"
-              onClick={modalControl.onOpen}
-            >
-              Change
-            </Link>
-          )}
-        </Stack>
-      ) : (
-        <Button onClick={modalControl.onOpen}>Record Video</Button>
-      )}
-    </>
-  )
-}
-
-export default VideoField
